requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.4.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
